// src/common/enums/Log.ts
/*
 * @Description:
 * @Author: Mary Qian
 * @Date: 2023-08-31 16:29:56
 * @LastEditTime: 2023-11-01 14:16:47
 * @LastEditors: Mary Qian
 */
/**
 * @description Enums of log
 * @author Shangmin Dou
 * @date 2021-05-15
 */
export enum Log {
    MOBILE_INFO = 'MOBILE INFO',
    MOBILE_WARN = 'MOBILE WARN',
    MOBILE_ERROR = 'MOBILE ERROR',
    NATIVE_LOG = 'NATIVE LOG'
}