// src/common/enums/NotificationType.ts
/*
 * @Description:
 * @Author: Mary Qian
 * @Date: 2023-11-09 14:54:18
 * @LastEditTime: 2023-11-09 15:10:38
 * @LastEditors: Mary Qian
 */

export enum NotificationType {
    UPDATE_MY_TEAM_BADGE = 'UpdateMyTeamBadge',
    UPDATE_MY_CUSTOMER_BADGE = 'UpdateMyCustomerBadge',
    RECALCULATE_MY_TEAM_AND_CUSTOMER_BADGE = 'RecalculateMyTeamAndCustomerBadge'
}