// src/savvy/enums/DeviceEvent.ts
/*
 * @Author: your name
 * @Date: 2021-10-19 02:58:20
 * @LastEditTime: 2021-11-02 01:34:09
 * @LastEditors: your name
 * @Description: In User Settings Edit
 * @FilePath: /Halo_Mobile/src/enums/DeviceEvent.ts
 */
export enum DeviceEvent {
    REFRESH_LEAD_LIST = 'refresh_lead_list',
    REFRESH_BADGE = 'refresh_badge',
    REFRESH_OPEN_LEAD_LIST = 'refresh_open_lead_list',
    START_SCHEDULE_MEETING = 'start_schedule_meeting',
    EDN_SCHEDULE_MEETING = 'end_schedule_meeting',
    CAROUSEL_REFRESH = 'carousel_refresh',
    IS_LANDSCAPE_MODE = 'is_landscape_mode'
}